requirements.txt
hypercorn==0.14.4
fastapi==0.115.0
uvicorn==0.30.1
python-docx==1.1.0
pypandoc==1.13
